backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
fastai==2.7.12
torch==2.0.1
pillow==9.5.0
numpy==1.24.3
werkzeug==2.3.7